// frontend/src/app/services/models/task-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TaskResponse {
  assignee?: number;
  category?: string;
  description?: string;
  duedate?: string;
  id?: number;
  importance?: 'HIGH' | 'MEDIUM' | 'LOW';
  status?: 'FINISHED' | 'IN_PROGRESS' | 'PENDING';
  title?: string;
}